weather-app/requirements.txt
Django==2.1.5
django-allauth==0.38.0
django-crispy-forms==1.7.2
djangorestframework==3.9.1
psycopg2-binary==2.7.6.1
pytz==2018.9
requests==2.21.0

